Remove shebang or make file executable
Almost all python files in the project have shebang, but they are not
intended to be executed directly so remove the shebang if the file is
not intended to be executed directly. Set executable bit to files which
are meant to be executed directly.

Accidentally this also fixes bunch of trailing white-spaces as my editor
does that for me - bonus :). No actual code change included!