Use python2 to build and run MysticMine.
PEP-394 defines that python2 should point to the system's python >= 2
installation, and python3 to the system's python >= 3.

Since many systems nowadays use python3 as the default python (ie: python is a
symlink to python3), python2 should explicitly be specified when python < 3 is
required.

This commit updates the build scripts, and MysticMine to use python2 by
default, since it's dependecies don't yet work on python3.